Extract printMdxJsxAttribute into a new file
